[framework] Remove code duplication
